Periods don't need to be stored
The stored periods aren't used anywhere, so we can get rid of them.
This also allows to compute the current period with simple arithmetic,
instead of in a loop.

The (new) implementation of `updatePeriod` is deliberately not using
SafeMath, because (a) the calculations are safe without, and (b) SafeMath
would significantly reduce readability of these longer expressions.
If, however, there is a SafeMath-all policy, this can be changed, of
course...