Fix bug in updateDelegateKey
In the pre-commit code it is, under very special circumstances,
possible to remove a well-behaving(*) party's delegateKey:

1. Assume Alice isn't a Moloch member or applicant yet,
   but intends to become one.
2. Malicious Bob knows Alice and can guess beforehand which address
   she'll use to apply. Let's call this address A.
3. Bob uses A as delegateKey, before Alice applies.
4. Alice becomes an applicant and finally a member.
   Her delegateKey is then initialized to A.
5. Now, if Bob updates his delegateKey, the delegateKey entry in
   Bob's member struct is still A, and it can be seen that the
   (pre-commit) code sets memberAddressByDelegateKey[A] to address(0),
   effectively removing Alice's delegateKey.

(*) "Well-behaving" here means: "not using an address they don't control".

The underlying problem is that the delegateKey entry in a member struct
can't always be trusted, so we must be careful not to overwrite data
based on this possibly outdated information.