Prevent SafeMath errors in processProposal
If a proposal with a ridiculously high number of requested shares
(close to the maximum uint) was accepted, then processProposal
would revert with a SafeMath error when adding the new shares to
the existing ones, and the tribute and deposit would be frozen in
the Moloch. Moreover, the processing of proposals would stop
entirely.
This commit introduces a fix that prevents overflows in the
existing calculations with shares.

Note: The last chunk of this commit is just to satisfy the apparent
SafeMath-everything policy.
In fact, it is one point of this commit that the multiplication
with the dilutionBound *should not* overflow, so the `mul` should
never revert.